app/backend/requirements.txt
Flask@2.2.2
openai@openai-0.27.4
azure-maps-search@1.0.0b2
azure-core@1.25.0
python-dotenv@1.0.0